Промежуточный комит, сделана половина боя (помогите боже есуси)

+++ cards_app/utils/common.py
@@ -4,11 +4,11 @@ from logging import getLogger
 logger = getLogger(__name__)
 
 
-def time_difference_check(check_time: datetime, need_hours: int) -> tuple[bool, float]:
+def time_difference_check(check_time: datetime, need_hours: int) -> tuple[bool, int]:
     """ Проверяет прошло ли необходимое количество часов для действия """
 
     difference = datetime.now() - check_time
-    hours = difference.total_seconds() // 3600
+    hours = int(difference.total_seconds() // 3600)
 
     return hours >= need_hours, hours
